getAvg

+++ src/test/TrainingTest.java
@@ -18,7 +18,7 @@ public class TrainingTest {
 		assertNotNull(t.grade);
 		Integer bestGrade = t.getBestGrade();
 		assertNotNull(bestGrade);
-		int avg = t.getAvg();
+		double avg = t.getAvg();
 		assertNotNull(avg);
 	}
